backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
google-cloud-aiplatform
google-auth
google-api-python-client
vertexai
diffusers
torch
transformers
pillow
